build: Add fixup script to fix package type
At first there was a problem in the ES Module compilation, because the
package.json at the root of the project was not set to "module" type.

This commit adds a script to fix this problem by setting the appropriate
package type in the package.json file at the root of each compilation.